% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/affine.decomposition.R
\name{affine.decomposition}
\alias{affine.decomposition}
\title{Affine subspace characterization by linear subspace and orthogonal vector.}
\usage{
affine.decomposition(subspace)
}
\arguments{
\item{subspace}{An k-by-n matrix of coordinates spanning the (k -
1)-dimensional affine subspace.}
}
\description{
Characterize an affine subspace by the parallel linear subspace and the
orthogonal vector from the origin.
}

