% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/affine.projection.R
\name{affine.projection}
\alias{affine.projection}
\title{Orthogonal projection to an affine subspace.}
\usage{
affine.projection(cloud, subspace)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{subspace}{An k-by-n matrix of coordinates spanning the (k - 1)-dimensional
affine subspace.}
}
\description{
This function calculates the orthogonal projection of a point (cloud) in a
Euclidean space to a affine subspace.
}

