% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/affine_recomposition.R
\name{affine_recomposition}
\alias{affine_recomposition}
\title{Affine subspace generation from a point and a linear subspace}
\usage{
affine_recomposition(point, subspace)
}
\arguments{
\item{point}{The n coordinates of a reference point.}

\item{subspace}{An k-by-n matrix (or n-coordinate vector when k equals 1) of coordinates spanning the k-dimensional linear subspace.}
}
\description{
This function returns the affine subspace through an inputed point and parallel (and equidimensional to) an inputed linear subspace.
}

