% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/axis_coordinates.R
\name{axis_coordinates}
\alias{axis_coordinates}
\title{Variables attached to an axis}
\usage{
axis_coordinates(cloud, dir_vec, variable)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional
space.}

\item{dir_vec}{An n-dimensional vector giving the direction of the axis
through the cloud.}

\item{variable}{Character; which variable to calculate, from among
\code{covariant}, \code{calibrated}, \code{standard}, and \code{axial}.
Does not default.}
}
\description{
The following functions inpute a point cloud (as a matrix of coordinates) and
an axis (as a direction vector), and return the coordinates of a desired
variable with respect to the cloud attached to the axis.
}

