% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/barycenter.R
\name{barycenter}
\alias{barycenter}
\title{Barycenter of a point cloud.}
\usage{
barycenter(cloud, weights)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{weights}{An m-element vector of point masses (weights).}
}
\description{
This function calculates the barycenter (weighted centroid) of a point cloud.
}

