% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cloud.decomposition.R
\name{cloud.decomposition}
\alias{cloud.decomposition}
\title{Fitted and residual clouds.}
\usage{
cloud.decomposition(cloud, subspace, resid = "cloud", ...)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{subspace}{An k-by-n matrix of coordinates spanning the k-dimensional linear
subspace or (k - 1)-dimensional affine subspace.}

\item{resid}{Whether to return the residuals as a cloud or as deviations.
Defaults to "cloud".}

\item{...}{Additional arguments passed to \code{\link{projection}}.}
}
\description{
Given a subspace, decompose a point cloud into its fitted cloud and either
its residual deviations or its residual cloud.
}

