% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cloud.variance.test.R
\name{cloud.variance.test}
\alias{cloud.variance.test}
\title{Test decomposition of cloud variance.}
\usage{
cloud.variance.test(cloud, weights, subspace, ...)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional
space.}

\item{weights}{An m-element vector of point masses (weights).}

\item{subspace}{An k-by-n matrix of coordinates spanning the k-dimensional
linear subspace or (k - 1)-dimensional affine subspace.}

\item{...}{Additional arguments passed to \code{\link{cloud.decomposition}}.}
}
\description{
This function verifies that the variance of a cloud is the sum of its
variance in the direction of any subspace and the variance of the residual
cloud with respect to that subspace.
}

