% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/direction.variance.R
\name{direction.variance}
\alias{direction.variance}
\title{Variance in the direction of a subspace.}
\usage{
direction.variance(cloud, weights, subspace, ...)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{weights}{An m-element vector of point masses (weights).}

\item{subspace}{An k-by-n matrix of coordinates spanning the k-dimensional subspace.}

\item{...}{Additional arguments passed to \code{\link{projection}}.}
}
\description{
This function calculates the variance of a point cloud in the direction of a
subspace.
}

