% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance between two points.}
\usage{
distance(x, y, ...)
}
\arguments{
\item{x}{The n coordinates of one point.}

\item{y}{The n coordinates of the other point.}
}
\description{
This function calculates the Euclidean distance between two points.
}

