% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{distances}
\title{Distances between pairs of points.}
\usage{
distances(x, y, ...)
}
\arguments{
\item{x}{An m-by-n matrix of the n coordinates of m first points.}

\item{y}{An m-by-n matrix of the n coordinates of m second points.}
}
\description{
This function calculates the Euclidean distances between pairs of points.
}

