% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eff.R
\name{eff}
\alias{eff}
\title{Effect linear mapping}
\usage{
eff(cloud, point, coord_vec)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional
space.}

\item{point}{An n-coordinate point in the same space as the cloud.}

\item{coord_vec}{An m-dimensional vector giving the coordinates in the space
of measures.}
}
\description{
This function encodes the effect linear mapping from the space of measures
over a cloud index set $J$ to the vector space containing the cloud $M^J$,
relative to a given point $P$.
}

