% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/huyghen.test.R
\name{huyghen.test}
\alias{huyghen.test}
\title{Test first Huyghen's theorem}
\usage{
huyghen.test(point, cloud, weights)
}
\arguments{
\item{point}{The n coordinates of a reference point.}

\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{weights}{An m-element vector of point masses (weights).}
}
\description{
This function calculates both sides of the equation of Huyghen's theorem.
}

