% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/huyghen_general_test.R
\name{huyghen_general_test}
\alias{huyghen_general_test}
\title{Test general Huyghen's theorem}
\usage{
huyghen_general_test(cloud, weights, subspace, ...)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional
space.}

\item{weights}{An m-element vector of point masses (weights).}

\item{subspace}{An k-by-n matrix of coordinates spanning the k-dimensional
affine subspace.}

\item{...}{Additional arguments (eventually) passed to \code{\link{projection}}.}
}
\description{
This function calculates both sides of the equation of Huyghen's theorem.
}

