% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inertia.R
\name{inertia}
\alias{inertia}
\title{Inertia of a point cloud with respect to a reference point.}
\usage{
inertia(point, cloud, weights, pointwise = FALSE, relative = FALSE)
}
\arguments{
\item{point}{The n coordinates of a reference point.}

\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{weights}{An m-element vector of point masses (weights).}

\item{pointwise}{Logical; whether to return pointwise (versus cloud)
inertias. Defaults to \code{FALSE}.}

\item{relative}{Logical; whether to return relative (versus absolute)
pointwise inertias. Defaults to \code{FALSE}.}
}
\description{
This function calculates the inertia of a point cloud with respect to a
reference point.
}

