% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linear.projection.R
\name{linear.projection}
\alias{linear.projection}
\title{Orthogonal projection to a linear subspace.}
\usage{
linear.projection(cloud, subspace)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{subspace}{An k-by-n matrix of coordinates spanning, together with the origin,
the h-dimensional linear subspace.}
}
\description{
This function calculates the orthogonal projection of a point (cloud) in a
Euclidean space to a linear subspace.
}

