% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projection.R
\name{projection}
\alias{projection}
\title{Orthogonal projection to a linear subspace.}
\usage{
projection(cloud, subspace, type = "affine", ...)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{subspace}{An k-by-n matrix of coordinates spanning the k-dimensional linear
subspace or (k - 1)-dimensional affine subspace.}

\item{type}{Whether to interpret h as an affine or a linear subspace.
Defaults to "affine".}
}
\description{
This function calculates the orthogonal projection of a point (cloud) in a
Euclidean space to a linear subspace.
}

