% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/projection.test.R
\name{projection.test}
\alias{projection.test}
\title{Test barycenter preservation under projection}
\usage{
projection.test(cloud, subspace, ...)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{subspace}{An k-by-n matrix of coordinates spanning the k-dimensional linear
subspace or (k - 1)-dimensional affine subspace.}

\item{...}{Additional arguments passed to \code{\link{projection}}.}
}
\description{
This function compares the barycenter of the projection of a point cloud to
the projection of the barycenter of the cloud.
}

