% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/residual.square.mean.R
\name{residual.square.mean}
\alias{residual.square.mean}
\title{The residual square mean of a subspace.}
\usage{
residual.square.mean(cloud, weights, subspace, ...)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{weights}{An m-element vector of point masses (weights).}

\item{subspace}{An k-by-n matrix of coordinates spanning the h-dimensional subspace.}

\item{...}{Additional arguments passed to \code{\link{projection}}.}
}
\description{
This function calculates the residual square mean of a point (cloud) with
respect to a subspace, i.e. the weighted mean of the squares of the distances
from the point(s) to the subspace.
}

