% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{Target}
\alias{Target}
\title{Target example.}
\format{A two-column matrix.}
\source{
Le Roux, Brigitte, Rouanet, Henry (2005) *Geometric Data Analysis*
\url{http://www.springer.com/us/book/9781402022357}
}
\description{
These data provide a running example of a Euclidean point cloud.
}

