% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tom.R
\name{tom}
\alias{tom}
\title{Measure space endomorphism}
\usage{
tom(cloud, point, coord_vec)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional
space.}

\item{point}{An n-coordinate point in the same space as the cloud.}

\item{coord_vec}{An m-dimensional vector giving the coordinates in the space
of measures.}
}
\description{
This function encodes the endomorphism $\mathcal{Tom}$ no the space of
measures over a cloud index set obtained by composing $\mathrm{Vac}$ with
$\mathrm{Eff}$.
}

