% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vac.R
\name{vac}
\alias{vac}
\title{Variable covariant linear mapping}
\usage{
vac(cloud, point, dir_vec)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional
space.}

\item{point}{An n-coordinate point in the same space as the cloud.}

\item{dir_vec}{An n-dimensional vector giving the direction of the axis
through the cloud.}
}
\description{
This function encodes a linear mapping that sends a vector in the vector
space containing a point cloud $M^J$ to the vector space of variables over
$J$, whose coordinates are the covariant coordinates of the cloud, relative
to a given point $P$, with respect to the axis generated by the vector.
}

