% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Variance of a point cloud.}
\usage{
variance(cloud, weights, ...)
}
\arguments{
\item{cloud}{An m-by-n matrix of coordinates for m points in n-dimensional space.}

\item{weights}{An m-element vector of point masses (weights).}

\item{...}{Additional arguments passed to \code{\link{inertia}}.}
}
\description{
This function calculates the variance of a point cloud.
}

